requirements.txt
asgiref==3.2.7
Django==3.0.4
pytz==2019.3
sqlparse==0.3.1
uWSGI==2.0.18
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
django-semantic-ui==1.2.3
